gencrypto: jose: cose: support ED25519 and ED448

+++ include/libwebsockets/lws-jwk.h
@@ -66,7 +66,7 @@ struct lws_jwk_parse_state {
 	int pos;
 	int cose_state;
 	int seen;
-	unsigned short possible;
+	unsigned int possible;
 };
 
 /** lws_jwk_import() - Create a JSON Web key from the textual representation
